match based on the tag for the package we are building
If you just find the most recent tag during a --test build
you run into git finding recent tags for *other* packages that are also
in the same git repo.  For example, if you have a git repo with 2
packages:

package-a
package-b

and you make a tag on package-a and have:

package-a-0.1.1-1

then you make a few commits and someone does a tag on package-b:

package-b-0.1.1-1

then a few more commits are done on code in package-a but nobody retags.
Now git describe will look back for the most recent tag and find
package-b-0.1.1-1 and have a commit count of the number of commits from
the package-b tag and not from the package-a tag.  By passing in the
--match option on 'git describe' we force git to only look for the tag
for the current package, not the most recent in the repo.