Add new mechanism for passing custom arguments to builders.
This is something to avoid using if possible, you wouldn't want to
require these for a builder that's the default for your project,
otherwise you'd need to specify it every time you ran the CLI.

However builders like the mock builder need to know which mock config to
use, and it could be multiple values for one project trying to do builds
for multiple OSs.

Adding a "builder.prop = value" to a releaser target section will pass
the valie prop to the builder configured.