use -rlvz instead of -avz for rsync.
We keep getting permission denied with trying to set time on directories.
According to the rsync man page:
-a, --archive               archive mode; equals -rlptgoD (no -H,-A,-X)

Hrm so what do those mean?

 -r, --recursive             recurse into directories
 -l, --links                 copy symlinks as symlinks
 -g, --group                 preserve group
 -o, --owner                 preserve owner (super-user only)
 -D                          same as --devices --specials
 -p, --perms                 preserve permissions
 -t, --times                 preserve modification times

So by having -a, that means -t automatically gets added which is not
what we want. And in some cases we were adding -no-p and -no-g
again added because -a includes them.

So by switching to -rlvz we avoid all these problems.