Add support for test build releases.
Adding "builder.test = 1" to a target in releasers.conf will allow you
to submit test builds out to a yum repository. This can be useful for
automated nightly builds. This is done as a builder arg rather than a
CLI argument to prevent anyone from accidentally doing a --test release
for something important. Instead it is a feature you configure once for
your targets.