Fix DistributionReleaser with GemBuilder
Due to error in GemBuilder.tgz method generating the list of sources
there was problem with adding the sources into lookaside (the list of
sources contained all the files in the directory, even directories).

Thanks to 9feb9cf48d0e24a74fefed8c41f8b2476882bac2 the tgz method in
NoTgzBuilder is generic enough to be able to extract the source files
even for GemBuilder.