Add more debugging facilities
The `debug` method now takes an optional 2nd argument, which
is a command to run via `run_command` method.

The `run_command` method now takes an optional 2nd argument, which
is used to always print output of command.

If --debug is passed on the CLI, run `git archive` separately
to make problems visible.