propagate docs to docker public registry
Add URL of upstream tito repo to HACKING, then
symlink HACKING as README.md in each Dockerfile dir.

Why?

If a README.md exists in same dir as Dockerfile (incl. symlink),
and it's set up as a
[trusted build](https://index.docker.io/help/docs/#trustedbuilds),
then index.docker.io automatically shows the readme on the info
page of the trusted build.