better release number for untagged packages
If package is unttaged and you build it with --test then previously 0 was used as number of commits
since last tag.
This results in packages which has no upgrade path.
With this patch it use number of commits since initial commit, which is growing number.
Therefore packages can be upgraded without usage of --force.