Fixup Fedora Dockerfiles to work correctly
The test suite tests both Python 2 and Python 3, but the spec file
doesn't install both sets of dependencies, so we need to declare
the Python 2 dependencies to be installed in the Dockerfiles.

In addition, there's no reason to drag in yum-utils, since
"dnf builddep" exists and serves the same purpose.