force VersionTagger when tito tag is called with --use-version
In 6b57b0b, the code that checked wither we used `--use-version` was
removed, because both `VersionTagger` and `ForceVersionTagger` are identical
now. However, this breaks setups where the default tagger is
`ReleaseTagger` but the user still wants to use `--use-version` sometimes.

Thus let's force `VersionTagger` when `--use-version` was given and only
use whatever is in `tito.props` when it was not.

Fixes: #274

Signed-off-by: Evgeni Golov <egolov@redhat.com>