Correct and make less confusing the conditional Python macros
This was using 'python_sitelib' in a non-standard way (it got
defined to either python2_sitelib or python3_sitelib depending
on whether we're doing a python2 or python3 build). Igor saw
the uses of 'python_sitelib' and assumed it was just using the
old name for 'the python 2 sitelib' and changed them all to
'python2_sitelib'...which broke all Fedora builds.

This fixes that problem and renames the package's confusing
'python_sitelib' and 'pythonbin' so hopefully no-one will make
the same mistake in future.

Signed-off-by: Adam Williamson <awilliam@redhat.com>