RPM spec fixes: deps, confusing macro names...
This patch brings across some fixes which have been made to the
RPM spec in the Fedora package. The %clean section should no
longer be necessary anywhere and is removed. The fedora-cert
dep is no longer needed, and is not available on Fedora Rawhide
(thanks to bcl for that one). The BuildRequire for 'python-devel'
is changed to 'python2-devel'. The require for 'python-cheetah'
is changed to '/usr/bin/cheetah', as in modern Fedora, the binary
is in python3-cheetah, not python2-cheetah. The cheetah code
should probably be changed to import now anyway, since python3-
cheetah now exists. The URL and Source0 locations are updated to
github instead of the old site, as the old site is now 404.

Finally, two confusing macro names are changed. The spec was using
a custom redefinition of 'python_sitelib', which is confusing; a
Fedora Python dev saw it and changed it to 'python2_sitelib',
assuming it was simply a use of the old name, but that actually
broke Fedora builds. To avoid that kind of confusion, let's
rename the custom '%python_sitelib' to '%our_sitelib', and rename
'%pythonbin' to '%ourpythonbin'.

Signed-off-by: Adam Williamson <awilliam@redhat.com>