Define _sourcedir for get_project_name RPM query
Follow-up for be89941
Fix: #311
Supersedes: #313

The `get_project_name` function is called in an early phase when
the temporary `SOURCES` directory for a current build doesn't exist yet.
Therefore as a little workaround, we set `_sourcedir` to an upstream
directory containing the package spec file.

Defining a custom `_sourcedir` (and therefore not using
`~/rpmbuild/SOURCES`) shouldn't be a problem because tito builds
should not depend on anything outside of the upstream git repository.

The use-case described in PR#313 works as expected.

    SOURCE1: somecool.macros
    %include %{SOURCE1}

For multi-spec projects, where you have e.g

    .
    ├── foo
    │   ├── foo.spec
    ├── bar
    │   ├── bar.spec

It is possible to include the common macros from a parent (or some
other) directory by having a symlink.

    cd foo
    ln -s ../somecool.macros ./

Or possibly some other way which I am currently not aware of
because I wasn't able to e.g. persuade `%include` to include from
the parent directory.