Fix copy_extra_sources for alternative builders
Fix #367

We accidentally introduced `copy_extra_sources` method as a part of
`BuilderBase` class. Which wouldn't be a problem, except for the fact
that its implementation uses properties, that are not a part of
that class but are rather defined in **some** of its descendats.

Let's define this method only abstractly and let more specific
builder classes to provide their own implementation if it makes
sense for them.