add Tito. The mascot.
This is bit of archeology.
Tito was a plush toy and mascot of Red Hat Network (RHN)_team.
Later when we opensourced Satellite as the Spacewalk project
we created this project. And by that time Tito was already missing
for some time. So in honor of our mascot we decided to name this
project Tito.